Lexer unit tests split into different tokens
